google-ads-scraper/requirements.txt
playwright==1.40.0
playwright-stealth==2.0.3
pandas==2.1.1
tldextract==3.6.0
requests==2.31.0
beautifulsoup4==4.12.3
email-validator==2.1.0
dnspython==2.6.1
httpx==0.27.0
